--- a/tests/conftest.py
+++ b/tests/conftest.py
@@ -413,6 +413,10 @@ def legacy_patchable_time():
         # Ensure point_in_time is UTC
         point_in_time = event.dt_util.as_utc(point_in_time)
 
+        # Since this is called once, we accept a HassJob so we can avoid
+        # having to figure out how to call the action every time its called.
+        job = action if isinstance(action, ha.HassJob) else ha.HassJob(action)
+
         @ha.callback
         def point_in_time_listener(event):
             """Listen for matching time_changed events."""
@@ -429,7 +433,7 @@ def legacy_patchable_time():
             setattr(point_in_time_listener, "run", True)
             async_unsub()
 
-            hass.async_run_job(action, now)
+            hass.async_run_hass_job(job, now)
 
         async_unsub = hass.bus.async_listen(EVENT_TIME_CHANGED, point_in_time_listener)
 
@@ -441,6 +445,8 @@ def legacy_patchable_time():
         hass, action, hour=None, minute=None, second=None, local=False
     ):
         """Add a listener that will fire if time matches a pattern."""
+
+        job = ha.HassJob(action)
         # We do not have to wrap the function with time pattern matching logic
         # if no pattern given
         if all(val is None for val in (hour, minute, second)):
@@ -448,7 +454,7 @@ def legacy_patchable_time():
             @ha.callback
             def time_change_listener(ev) -> None:
                 """Fire every time event that comes in."""
-                hass.async_run_job(action, ev.data[ATTR_NOW])
+                hass.async_run_hass_job(job, ev.data[ATTR_NOW])
 
             return hass.bus.async_listen(EVENT_TIME_CHANGED, time_change_listener)
 
@@ -485,8 +491,8 @@ def legacy_patchable_time():
             last_now = now
 
             if next_time <= now:
-                hass.async_run_job(
-                    action, event.dt_util.as_local(now) if local else now
+                hass.async_run_hass_job(
+                    job, event.dt_util.as_local(now) if local else now
                 )
                 calculate_next(now + datetime.timedelta(seconds=1))